% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement.R
\name{checkdependencies}
\alias{checkdependencies}
\title{Check if Dependencies are installed}
\usage{
checkdependencies()
}
\description{
A function to check if dependencies are installed. Installed dependncies must be loaded manually.
}
