% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicpipeline.R
\name{deg.acrossclusters}
\alias{deg.acrossclusters}
\title{Differential expression testing for clusters in a single sample}
\usage{
deg.acrossclusters(
  sobj,
  idents = NULL,
  test = NULL,
  latent.vars = NULL,
  outdir = NULL
)
}
\arguments{
\item{sobj}{A seurat object, usually one on which Seurat::FindClusters() has been run.}

\item{idents}{A string referring to a categorical column of the seurat metadata, such as the output of Seurat::FindClusters(). Defaults to 'seurat_project'.}

\item{test}{A string connoting which DEG test to perform. See ?Seurat::FindMarkers() for details on options. Defaults to "MAST".}

\item{latent.vars}{A string or character vector referring to which columns to use a "latent variables", which some of the tests will attempt to regress out the effect of. Some tests (such as MAST) can only perform this for quantitative, continuous variables. Defaults to ('nFeature_RNA', 'nCount_RNA', 'percent.mito').}

\item{outdir}{a string connoting which directory to save results in. Will create directory if it does not exist. Will not overwrite. Defaults to the following: "ClusterDEG_(Seurat object project name)_(date)_(test)"}
}
\value{
Saves output files to outdir. Will not return anything else besides printing progress reports to the standard out.
}
\description{
A function to run differentially expressed gene (DEG) testing and save to disk the output in a format compatible with gseapipeline.clusters() function.
This function takes a while. To help deal with the risk of interruptions, this function is written in a way that tries to pick up where previously left off if stopped part-way through.
}
\examples{
\dontrun{
pdf('qcplots.pdf')
deg.acrossclusters(sobj = sobj)
dev.off()
}
}
