% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone_gsea.R
\name{gsea.clusters}
\alias{gsea.clusters}
\title{GSEA for clusters.}
\usage{
gsea.clusters(
  degdf,
  clustercolname = NULL,
  pathways = NULL,
  nperm = NULL,
  weightmethod = NULL,
  onlypos = NULL,
  filter_nonsig_pathways = NULL
)
}
\arguments{
\item{degdf}{A dataframe in the format of the output of Seurat::FindAllMarkers.}

\item{pathways}{A named "list of lists" of genes. The format of this object is important for functionality of this function and can be previewed in the tamlabscpipeline::hallmark object. The names of each list element provides the Y axis; the genes within each list element are the target genes used for GSEA. If NULL, defaults to the Msigdb's Hallmark pathways for mouse.}

\item{nperm}{An integer denoting how many permutations FGSEA will run. Default = 10000.}

\item{weightmethod}{A string, one of either "pvalue", "padj", or "foldchange".}

\item{onlypos}{T/F. Whether to filter each cluster to only include upregulated genes.}

\item{filter_nonsig_pathways}{T/F. Whether to remove non-significant rows from resulting heatmap. Useful if running very large numbers of pathways, or to generate finalized plots after exploratory analysis. Default = F.}
}
\value{
Prints a heatmap to the standard out, or to a pdf.
}
\description{
A function to perform GSEA on differentially expressed gene lists from clusters. Expects column names consistent with output of Seurat::FindAllMarkers(). Essentially, a dataframe where each row corresponds to a gene, and columns consists of the following: "p_val", "avg_logFC", "pct.1", "pct.2", "p_val_adj", "cluster", and "gene". Minimally requires "cluster", "gene", and "avg_logFC". Pvalue and/or adjusted pvalue can be included if weighting by those values.
Output plots can be difficult to read if many pathways are used. Recommended 50-75 or so pathways as the maximum, many more will generate very ugly heatmaps.
Utilizes special characters, so pdf printing may be problematic. Alternative PDF devices should be able to handle this though, one can use Quartz if on Mac, or CairoPDF if on Windows.
}
\examples{
\dontrun{
sobjmarkers <- Seurat::FindAllMarkers(sobj)
gsea.clusters(sobjmarkers)
}
}
