% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicpipeline.R
\name{gseapipeline.clusters}
\alias{gseapipeline.clusters}
\title{A function for GSEA using the output of differential gene expression testing}
\usage{
gseapipeline.clusters(
  inputfolder,
  pathways,
  nperm = NULL,
  weightmethod = NULL,
  makepdf = NULL,
  pdfname = NULL,
  filter_nonsig_pathways = NULL
)
}
\description{
A function for GSEA using the output of differential gene expression testing
}
