% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicpipeline.R
\name{seuratpipeline}
\alias{seuratpipeline}
\title{Basic SC Pipeline for QC and clustering}
\usage{
seuratpipeline(
  data,
  format,
  transcript_gene_file,
  project = NULL,
  baselinefilter.mad = NULL,
  baseline.mito.filter = NULL,
  madmax.dist.percentmito.baseline = NULL,
  baseline.libsize.filter = NULL,
  madmax.dist.nCount_RNA.baseline = NULL,
  removemitomaxclust = NULL,
  iterativefilter = NULL,
  iterativefilter.libsize = NULL,
  iterativefilter.libsize.twosided = NULL,
  iterativefilter.libsize.lefttail = NULL,
  iterativefilter.mito = NULL,
  cellcycleregression = NULL,
  PCAgenelist = NULL,
  jackstraw = NULL,
  dims = NULL,
  res = NULL
)
}
\arguments{
\item{data}{a string containing a filepath with format connoted by the format parameter.}

\item{format}{a string, either 'dir' for cellranger dir output; 'h5' for cellranger h5 output, or 'kallisto' for the kallisto|bustools pipeline output}

\item{transcript_gene_file}{a string containing a filepath for the "transcript_gene" conversion file. Only used for Kallisto|bustools workflow.}

\item{baselinefilter.mad}{T/F; whether to perform "global" QC, ie without pre-clustering. Default = False.}

\item{baseline.mito.filter}{T/F; whether to perform global maximum mitochondrial content filtration using median absolute deviation threshold; will only work if baselinefilter.mad is set to True. Default is True.}

\item{madmax.dist.percentmito.baseline}{a numeric, or a string reading 'predict'. If numeric is provided, will use this as median absolute deviation threshold for global mito cutoff. If set to 'predict', will attempt to learn cutoff from data. Default = 'predict'}

\item{baseline.libsize.filter}{T/F; whether to perform global minimal lib size filtration using median absolute deviation threshold; will only work if baselinefilter.mad is set to True. Default is True.}

\item{madmax.dist.nCount_RNA.baseline}{a numeric, or a string reading 'predict'. If numeric is provided, will use this as median absolute deviation threshold for global libsize cutoff. If set to 'predict', will attempt to learn cutoff from data. Default = 'predict'}

\item{removemitomaxclust}{T/F ; whether to identify and remove abnormally high mitochondrial content clusters after first-pass clustering; if no baseline mito filtration is used, there will almost certainly be a mito-driven cluster. Identification is via Grubbs' test for outliers, based on Lukasz Komsta's implementation in the outliers package. Default is True.}

\item{iterativefilter}{T/F ; whether to perform iterative filtering on first-pass filtering. Default is True.}

\item{iterativefilter.libsize}{either one of two strings ('twosided' or 'lefttail') or False. Twosided will attempt to learn median abs. dev. cutoffs for both min and max to catch debris and, ostensibly, doublets. Lefttail will attempt to learn median abs. dev. threshold for min to catch debris; doublets may not accurately be captured by max cutoffs as this is more an artifact of sequencing than cell suspension. False will skip. Default is 'lefttail'.}

\item{iterativefilter.mito}{T/F; whether to learn right-tail median abs. dev. thresholds and filter maximal mitochondrial content from each cluster. May incorrectly remove mito-okay cells while missing true mito-hi cells. Default = F.}

\item{cellcycleregression}{either one of two strings ('total', 'difference'), or False. If false, still calculates cell cycle score but does not attempt correction. Default is False. See here: https://satijalab.org/seurat/v3.1/cell_cycle_vignette.html}

\item{PCAgenelist}{a character vector of gene names to use for PCA. If null, defaults to highly variable genes called by SCT. default is NULL.}

\item{jackstraw}{T/F; whether to perform jackstraw to score significant PCs for use in clustering / dimreduction. May be incompatible with SCT. Default is false.}

\item{dims}{an integer range. Controls graph construction prior to clustering and dimensionality reduction for visualization. Connotes which PC dimensions to use in clustering. Defaults to 1:30.}

\item{res}{a numeric, vector of numerics, or range of numerics. Controls assignment of cells to clusters. Connotes the "resolution" paramter used as correction in Louvain clustering. Defaults to c(0.5, 1.0, 1.5).}
}
\value{
will return a bunch of plots related to QC and an output in the form of a Seurat object to the standard out.
}
\description{
Basic SC pipeline for a single input dataset.
Perform QC and clustering of an input dataset.
}
\examples{
\dontrun{
pdf('qcplots.pdf')
sobj <- seuratpipeline('datafilepath.h5', format=h5)
dev.off()
}
}
