% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement.R
\name{grubbs}
\alias{grubbs}
\title{a test for outliers assuming normal distribution.
based on the grubbs.test() function from the outliers pkg by Lukasz Komsta}
\usage{
grubbs(x, type = 10, opposite = FALSE, two.sided = FALSE)
}
\description{
a test for outliers assuming normal distribution.
based on the grubbs.test() function from the outliers pkg by Lukasz Komsta
}
